+++ lib/tpl.php
@@ -8,4 +8,28 @@ class egr {
 		$path = kirby()->roots()->plugins() . DS . 'field-engineer' . DS . 'snippets' . DS . $name . '.php';
 		return tpl::load($path, $data);
 	}
+
+	public static function grid($field) {
+		$grid = '';
+		if(isset($field['rowWidth'])) {
+			$grid = ' egr-grid-item egr-grid-item-' . str_replace('/', '-', $field['rowWidth']);
+		}
+		return $grid;
+	}
+
+	public static function buttons($set) {
+		$buttons = array();
+		if(isset($set['buttons'])) {
+			$buttons = $set['buttons'];
+		}
+		return $buttons;
+	}
+
+	public static function count($field) {
+		$count = 0;
+		if(isset($field['sets'])) {
+			$count = count($field['sets']);
+		}
+		return $count;
+	}
 }