+++ snippets/template.php
@@ -7,7 +7,10 @@
 	</style>
 <?php endif; ?>
 <div class="egr">
-	<div class="egr-outline"><?php echo egr::snippet('outline', $args['outline']); ?></div>
+	<div class="egr-outline">
+		<?php echo egr::snippet('delete'); ?>
+		<?php echo egr::snippet('outline', $args['outline']); ?>
+	</div>
 	<div class="egr-presentation"><?php echo egr::snippet('presentation', $args['presentation']); ?></div>
 	<div class="egr-output">
 		<textarea